requirements.txt
python-dotenv
azure-cognitiveservices-speech
google-ai-generativelanguage
elevenlabs
pyaudio